+++ NEWS.md
@@ -1,4 +1,10 @@
-# Reactome 1.34.0
+# ReactomePA 1.41.0.991
+
++ support passing a GSON object to `enrichPathway(organism)` and `gsePathway(organism)`(2022-7-13, Wed)
++ add function `gson_Reactome` (2022-7-13, Wed)
+
+
+# ReactomePA 1.34.0
 
 + Bioconductor 3.12 release (2020-10-28, Wed)